Split code up into components
